# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TukAPI
  version: 1.0.0
  description: >
    OpenAPI draft for the Sri Lanka Police tuk-tuk tracking and movement logging API.
    This document covers authentication, protected master-data management, device ping
    ingestion, and live or historical location lookup endpoints.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /tuk-tuks:
    get:
      tags: [TukTuks]
      summary: List tuk-tuks
      operationId: listTukTuks
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ProvinceId'
        - $ref: '#/components/parameters/DistrictId'
        - $ref: '#/components/parameters/StationId'
        - $ref: '#/components/parameters/DriverId'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/TukTukStatus'
        - $ref: '#/components/parameters/PlateNumber'
        - $ref: '#/components/parameters/RegistrationNumber'
        - $ref: '#/components/parameters/TukTukSortBy'
        - $ref: '#/components/parameters/SortOrderDesc'
      responses:
        '200':
          description: Tuk-tuk list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TukTukListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [TukTuks]
      summary: Create a tuk-tuk
      operationId: createTukTuk
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTukTukRequest'
            example:
              registrationNumber: REG-2026-000145
              plateNumber: WP CAB-4567
              model: Bajaj RE
              color: Green
              provinceId: cm9x7a1bc0001abc123456789
              districtId: cm9x7a1bc0002abc123456789
              stationId: cm9x7a1bc0003abc123456789
              driverId: cm9x7a1bc0004abc123456789
              deviceId: cm9x7a1bc0005abc123456789
              status: ACTIVE
              notes: Demonstration vehicle for Colombo district.
      responses:
        '201':
          description: Tuk-tuk created successfully
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TukTukSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tuk-tuks/{id}:
    get:
      tags: [TukTuks]
      summary: Get a single tuk-tuk
      operationId: getTukTukById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Tuk-tuk returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TukTukSingleResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [TukTuks]
      summary: Partially update a tuk-tuk
      operationId: updateTukTuk
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTukTukRequest'
      responses:
        '200':
          description: Tuk-tuk updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TukTukSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tuk-tuks/{id}/location-history:
    get:
      tags: [TukTuks, Locations]
      summary: Query movement history for a single tuk-tuk
      operationId: getTukTukLocationHistory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/FromTime'
        - $ref: '#/components/parameters/ToTime'
        - $ref: '#/components/parameters/IgnitionOn'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/LocationHistorySortBy'
        - $ref: '#/components/parameters/SortOrderDesc'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Tuk-tuk location history returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationPingListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users:
    get:
      tags: [Users]
      summary: List authorized users
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ProvinceId'
        - $ref: '#/components/parameters/DistrictId'
        - $ref: '#/components/parameters/StationId'
        - $ref: '#/components/parameters/UserRole'
        - $ref: '#/components/parameters/IsActive'
        - $ref: '#/components/parameters/UserSortBy'
        - $ref: '#/components/parameters/SortOrderDesc'
      responses:
        '200':
          description: User list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Users]
      summary: Create an authorized law-enforcement user
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get a single authorized user
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Users]
      summary: Partially update an authorized user
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated law-enforcement users.
    DeviceTokenAuth:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Device token used for telemetry ingestion routes.
  headers:
    ETag:
      description: Weak entity tag for conditional GET support.
      schema:
        type: string
      example: W/"S2JkQm5GQ1hPVEp5R0FxQ0pmVQ"
    LastModified:
      description: Last modification timestamp for conditional GET support.
      schema:
        type: string
        format: date-time
      example: '2026-04-29T06:25:13.000Z'
    CacheControl:
      description: Cache directive for the returned representation.
      schema:
        type: string
      example: private, max-age=30, must-revalidate
    Location:
      description: Relative URI of the created resource.
      schema:
        type: string
      example: /api/v1/tuk-tuks/cm9x7a1bc0006abc123456789
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier.
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of records per page.
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search term.
    ProvinceId:
      name: provinceId
      in: query
      schema:
        type: string
      description: Filter by province ID.
...
    DistrictId:
      name: districtId
      in: query
      schema:
        type: string
      description: Filter by district ID.
    StationId:
      name: stationId
      in: query
      schema:
        type: string
      description: Filter by station ID.
    DriverId:
      name: driverId
      in: query
      schema:
        type: string
      description: Filter by driver ID.
    DeviceId:
      name: deviceId
      in: query
      schema:
        type: string
      description: Filter by device ID.
...
    IsActive:
      name: isActive
      in: query
      schema:
        type: boolean
      description: Filter by active status.
...
    PlateNumber:
      name: plateNumber
      in: query
      schema:
        type: string
      description: Filter by exact tuk-tuk plate number.
    RegistrationNumber:
      name: registrationNumber
      in: query
      schema:
        type: string
      description: Filter by exact tuk-tuk registration number.
...
    UserRole:
      name: role
      in: query
      schema:
        $ref: '#/components/schemas/UserRole'
      description: Filter by user role.
    TukTukStatus:
      name: status
      in: query
      schema:
        $ref: '#/components/schemas/TukTukStatus'
      description: Filter by tuk-tuk status.
...
    TukTukSortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [createdAt, updatedAt, registrationNumber, status]
        default: createdAt
      description: Tuk-tuk sort field.
...
    UserSortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [createdAt, updatedAt, fullName, email, role]
        default: createdAt
      description: User sort field.
...
    LocationHistorySortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [recordedAt, receivedAt, createdAt]
        default: recordedAt
      description: History sort field.
...
    SortOrderDesc:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort direction.
...
    FromTime:
      name: from
      in: query
      schema:
        type: string
        format: date-time
      description: Inclusive start of the requested time range.
    ToTime:
      name: to
      in: query
      schema:
        type: string
        format: date-time
      description: Inclusive end of the requested time range.
    IgnitionOn:
      name: ignitionOn
      in: query
      schema:
        type: boolean
      description: Filter by ignition state.
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema:
        type: string
      description: Conditional GET validator based on the last returned ETag.
    IfModifiedSince:
      name: If-Modified-Since
      in: header
      schema:
        type: string
        format: date-time
      description: Conditional GET validator based on resource modification time.
  responses:
    NotModified:
      description: Resource has not changed since the supplied cache validator.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Last-Modified:
          $ref: '#/components/headers/LastModified'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
    BadRequest:
      description: Request failed validation or query parsing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Validation failed.
            errors:
              - path: query.limit
                message: Number must be less than or equal to 100.
    Unauthorized:
      description: Authentication failed or credentials are missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication token is required.
    Forbidden:
      description: Authenticated caller does not have access to this resource or scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You are not authorized to access this resource.
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Requested resource was not found.
    Conflict:
      description: Request would violate a unique constraint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: A unique constraint would be violated by this request.
    UnprocessableEntity:
      description: Request is valid JSON but violates a business rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Tracking device is already assigned to another tuk-tuk.
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Too many requests, please try again later.
    InternalServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Internal server error.
  schemas:
    UserRole:
      type: string
      enum:
        - HQ_ADMIN
        - PROVINCIAL_ADMIN
        - DISTRICT_OFFICER
        - STATION_OFFICER
        - DEVICE_CLIENT
    TukTukStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - MAINTENANCE
        - IMPOUNDED
    DeviceStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - MAINTENANCE
        - RETIRED
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 2
      required: [page, limit, totalItems, totalPages]
    ErrorItem:
      type: object
      properties:
        path:
          type: string
          example: body.provinceId
        message:
          type: string
          example: Invalid input.
      required: [path, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
      required: [success, message]
...
    ProvinceSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
      required: [id, name, code]
    DistrictSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
      required: [id, name, code]
    StationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
      required: [id, name, code]
    DriverSummary:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        licenseNumber:
          type: string
        isActive:
          type: boolean
      required: [id, fullName, licenseNumber, isActive]
    DeviceSummary:
      type: object
      properties:
        id:
          type: string
        serialNumber:
          type: string
        status:
          $ref: '#/components/schemas/DeviceStatus'
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
      required: [id, serialNumber, status]
...
    TukTuk:
      type: object
      properties:
        id:
          type: string
        provinceId:
          type: string
          nullable: true
        districtId:
          type: string
          nullable: true
        stationId:
          type: string
          nullable: true
        driverId:
          type: string
          nullable: true
        deviceId:
          type: string
          nullable: true
        registrationNumber:
          type: string
        plateNumber:
          type: string
        model:
          type: string
        color:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TukTukStatus'
        notes:
          type: string
          nullable: true
        province:
          allOf:
            - $ref: '#/components/schemas/ProvinceSummary'
          nullable: true
        district:
          allOf:
            - $ref: '#/components/schemas/DistrictSummary'
          nullable: true
        station:
          allOf:
            - $ref: '#/components/schemas/StationSummary'
          nullable: true
        driver:
          allOf:
            - $ref: '#/components/schemas/DriverSummary'
          nullable: true
        device:
          allOf:
            - $ref: '#/components/schemas/DeviceSummary'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - registrationNumber
        - plateNumber
        - model
        - status
        - createdAt
        - updatedAt
...
    User:
      type: object
      properties:
        id:
          type: string
        provinceId:
          type: string
          nullable: true
        districtId:
          type: string
          nullable: true
        stationId:
          type: string
          nullable: true
        fullName:
          type: string
        email:
          type: string
          format: email
        badgeNumber:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        phoneNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        province:
          allOf:
            - $ref: '#/components/schemas/ProvinceSummary'
          nullable: true
        district:
          allOf:
            - $ref: '#/components/schemas/DistrictSummary'
          nullable: true
        station:
          allOf:
            - $ref: '#/components/schemas/StationSummary'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
        - updatedAt
    LocationPing:
      type: object
      properties:
        id:
          type: string
        tukTukId:
          type: string
        deviceId:
          type: string
        provinceId:
          type: string
        districtId:
          type: string
        stationId:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        speedKmh:
          type: number
          format: double
          nullable: true
        heading:
          type: number
          format: double
          nullable: true
        accuracyM:
          type: number
          format: double
          nullable: true
        ignitionOn:
          type: boolean
          nullable: true
        recordedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        tukTuk:
          $ref: '#/components/schemas/TukTuk'
        device:
          allOf:
            - $ref: '#/components/schemas/DeviceSummary'
          nullable: true
        province:
          $ref: '#/components/schemas/ProvinceSummary'
        district:
          $ref: '#/components/schemas/DistrictSummary'
        station:
          allOf:
            - $ref: '#/components/schemas/StationSummary'
          nullable: true
      required:
        - id
        - tukTukId
        - provinceId
        - districtId
        - latitude
        - longitude
        - recordedAt
        - receivedAt
...
    TukTukSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Tuk-tuk fetched successfully.
        data:
          $ref: '#/components/schemas/TukTuk'
      required: [success, message, data]
    TukTukListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Tuk-tuks fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/TukTuk'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
...
    UserSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: User fetched successfully.
        data:
          $ref: '#/components/schemas/User'
      required: [success, message, data]
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Users fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
...
    LocationPingListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Location history fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/LocationPing'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
...
    CreateTukTukRequest:
      type: object
      properties:
        registrationNumber:
          type: string
        plateNumber:
          type: string
        color:
          type: string
        model:
          type: string
        provinceId:
          type: string
        districtId:
          type: string
        stationId:
          type: string
        driverId:
          type: string
          nullable: true
        deviceId:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TukTukStatus'
        notes:
          type: string
      required:
        - registrationNumber
        - plateNumber
        - model
        - provinceId
        - districtId
        - stationId
    UpdateTukTukRequest:
      type: object
      properties:
        registrationNumber:
          type: string
        plateNumber:
          type: string
        color:
          type: string
        model:
          type: string
        provinceId:
          type: string
        districtId:
          type: string
        stationId:
          type: string
        driverId:
          type: string
          nullable: true
        deviceId:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TukTukStatus'
        notes:
          type: string
...
    CreateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        badgeNumber:
          type: string
        password:
          type: string
          format: password
        phoneNumber:
          type: string
        role:
          type: string
          enum:
            - HQ_ADMIN
            - PROVINCIAL_ADMIN
            - DISTRICT_OFFICER
            - STATION_OFFICER
        provinceId:
          type: string
        districtId:
          type: string
        stationId:
          type: string
        isActive:
          type: boolean
          default: true
      required: [fullName, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        badgeNumber:
          type: string
        password:
          type: string
          format: password
        phoneNumber:
          type: string
        role:
          type: string
          enum:
            - HQ_ADMIN
            - PROVINCIAL_ADMIN
            - DISTRICT_OFFICER
            - STATION_OFFICER
        provinceId:
          type: string
          nullable: true
        districtId:
          type: string
          nullable: true
        stationId:
          type: string
          nullable: true
        isActive:
          type: boolean
...